Demacrofy ADC
Split out common adc parts to avoid safety issues

Clean up adc

Update tests and examples for adc changes

Use indices for arrayified fields
