+++ Cargo.toml
@@ -89,13 +89,13 @@ rt = ["stm32g4/rt"]
 usb = ["dep:stm32-usbd"]
 stm32g431 = ["stm32g4/stm32g431"]
 stm32g441 = ["stm32g4/stm32g441"]
-stm32g471 = ["stm32g4/stm32g471"]
-stm32g473 = ["stm32g4/stm32g473"]
-stm32g474 = ["stm32g4/stm32g474"]
-stm32g483 = ["stm32g4/stm32g483"]
-stm32g484 = ["stm32g4/stm32g484"]
-stm32g491 = ["stm32g4/stm32g491"]
-stm32g4a1 = ["stm32g4/stm32g4a1"]
+stm32g471 = ["stm32g4/stm32g471", "adc3"]
+stm32g473 = ["stm32g4/stm32g473", "adc3", "adc4", "adc5"]
+stm32g474 = ["stm32g4/stm32g474", "adc3", "adc4", "adc5"]
+stm32g483 = ["stm32g4/stm32g483", "adc3", "adc4", "adc5"]
+stm32g484 = ["stm32g4/stm32g484", "adc3", "adc4", "adc5"]
+stm32g491 = ["stm32g4/stm32g491", "adc3"]
+stm32g4a1 = ["stm32g4/stm32g4a1", "adc3"]
 log-itm = ["cortex-m-log/itm"]
 log-rtt = []
 log-semihost = ["cortex-m-log/semihosting"]
@@ -107,6 +107,9 @@ defmt = [
     "embedded-io/defmt-03",
 ]
 cordic = ["dep:fixed"]
+adc3 = []
+adc4 = []
+adc5 = []
 
 [profile.dev]
 codegen-units = 1
